updated: HomeScreen components
