+++ core/design-system/src/main/java/dev/enesky/core/design_system/common/Placeholder.kt
@@ -31,6 +31,6 @@ fun Modifier.placeholder(color: Color) = composed {
     placeholder(
         visible = true,
         color = color,
-        shape = DoodleTheme.shapes.medium,
+        shape = DoodleTheme.shapes.small,
     )
 }
